Add status.go
Set the status only once. This allows us to (1) set some
error status but then (2) still render a normal page.

For example this is useful when someone forgets to input
anything into the post field and submits. We now have the
groundwork to just forward to the referal URL with an
error flash instead of an ugly full-screen error message.